refactor: 타 페이지와 동일한 레벨로 주문 페이지 컴포넌트 추상화
